third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGFEPointLightElement-dom-y-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'y' attribute of the SVGFEPointLightElement object.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS pointLight.getAttribute('y') is "120"
PASS pointLight.getAttribute('y') is "180"
PASS successfullyParsed is true

TEST COMPLETE
